refactor: redirect public packages from internal/sys to sys
Public packages (sys/mutex, window, overload/bbr) were importing
internal/sys/* duplicates instead of the public sys/* versions.
Redirected all imports to sys/* to eliminate the dependency on the
duplicate internal/sys packages.

This is a step toward removing internal/sys/* entirely once internal
packages are also migrated.
