+++ sys/mutex/recursive_mutex.go
@@ -5,7 +5,7 @@ import (
 	"sync"
 	"sync/atomic"
 
-	"github.com/songzhibin97/gkit/internal/sys/goid"
+	"github.com/songzhibin97/gkit/sys/goid"
 )
 
 // RecursiveMutex 包装一个Mutex,实现可重入
