+++ window/array.go
@@ -5,7 +5,7 @@ import (
 	"unsafe"
 
 	"github.com/songzhibin97/gkit/internal/clock"
-	"github.com/songzhibin97/gkit/internal/sys/safe"
+	"github.com/songzhibin97/gkit/sys/safe"
 )
 
 // AtomicArray 封装原子操作, 底层维护 []*Bucket
